feat: moved API route for user signups to /auth/register (and some small fixes) (#136)
