+++ pages/api/frontend/v0.1/users.ts
@@ -1,8 +1,7 @@
-import type { NextApiRequest, NextApiResponse } from 'next';
 import { PrismaClient } from '@prisma/client';
-import { hashAndSaltPassword, validatePassword, generateToken, sendTokenPerMail } from '../../../../util/auth';
-import { getSession } from 'next-auth/react';
 import { StatusCodes } from 'http-status-codes';
+import type { NextApiRequest, NextApiResponse } from 'next';
+import { getSession } from 'next-auth/react';
 
 const prisma = new PrismaClient()
 
@@ -10,86 +9,7 @@ export default async function handler(
     req: NextApiRequest,
     res: NextApiResponse
 ) {
-    // TODO: remove duplicate code, e.g. `const session = await getSession({ ... })`
-    
     switch(req.method) {
-        // TODO: instead of creating user via POST /users, it might be clearer for maintenance to use POST /auth/register
-        case 'POST':
-            const data = req.body;
-
-            const { email, password, firstName, lastName } = data;
-
-            // TODO: we should consider proxying all the environment flags with a configuration class
-            if (process.env.SIGNUPS_ENABLED === "false") {
-                res
-                    .status(StatusCodes.METHOD_NOT_ALLOWED)
-                    .json({ message: 'Not allowed - signups are currently disabled!'});
-                return;
-            }
-
-            if (!email || !email.includes('@')) {
-                res
-                    .status(StatusCodes.UNPROCESSABLE_ENTITY)
-                    .json({ message: 'Invalid data - email not valid'});
-                    return;
-            }
-
-            if (!(await validatePassword(password))) {
-                res
-                    .status(StatusCodes.UNPROCESSABLE_ENTITY)
-                    .json({ message: 'Invalid data - password consists of less than 8 characters'});
-                return;
-            }
-
-            const lookupUser = await prisma.user.findFirst({
-                where: {
-                    email: email,
-                    NOT: {
-                        isDeleted: true,
-                    }
-                }
-            });
-            
-            if (lookupUser) {
-                res
-                    .status(StatusCodes.CONFLICT)
-                    .json({ message: 'Conflict - email already in use'});
-                return;
-            }
-
-            const { hashedSaltedPassword, salt } = await hashAndSaltPassword(password);
-
-            const createdUser = await prisma.user.create({
-                data: {
-                    email: email,
-                    password: hashedSaltedPassword,
-                    salt: salt,
-                    firstName: firstName,
-                    lastName: lastName,
-                    isVerified: false,
-                }
-            });
-
-            const generatedToken = generateToken();
-        
-            var expiryDate = new Date();
-            // set expiryDate one week from now
-            expiryDate.setDate(expiryDate.getDate() + 7);
-    
-            const verificationToken = await prisma.verificationToken.create({
-                data: {
-                    userId: createdUser.id,
-                    token: generatedToken,
-                    expiryDate: expiryDate,
-                    isArchived: false,
-                }
-            });
-    
-            sendTokenPerMail(createdUser.email as string, createdUser.firstName as string, verificationToken.token, "VERIFY", "");
-            
-            res.status(StatusCodes.CREATED).json(email);
-            break;
-
         case 'GET':
             const session = await getSession({ req: req });
