+++ styles/Home.module.css
@@ -1,0 +1,12 @@
+.main {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 6rem;
+    min-height: 100vh;
+  }
+  
+  .main h1 {
+    margin-bottom: 64px;
+    margin-top: 64px;
+  }
